Dynamic recovery of missed SRDY interrupts
When NPI Server is compiled to use SRDY interrupts there
is a chance that interrupts are missed. This is mitigated by
capturing the perror and enable timeouts. The timeout is set
dynamically to balance system resource usage and performance.

Also added possibility to configure UART baudrate.

Added new compiler for BeagleBone Black. Link to toolchain is
found next to the compiler argument in the makefile.

Minor change to error management. Instead of a huge switch case
we now use a bit to indicate severity of error, and whether or
not the error requires reset.

+++ Projects/tools/LinuxHost/application/common/timer.h
@@ -66,7 +66,7 @@ typedef struct
 
 extern int    timer_init			(uint16 numOfThreads);
 
-extern uint8  timer_start_timerEx	(uint8 threadId, uint16 event, uint16 timeout);
+extern uint8  timer_start_timerEx	(uint8 threadId, uint16 event, uint32 timeout);
 
 extern uint8  timer_set_event		(uint8 threadId, uint16 event);
 extern uint8  timer_clear_event		(uint8 threadId, uint16 event);
